add ksp processor to bignumber and rename package
