+++ kbignum/src/commonMain/kotlin/korlibs/bignumber/BigExt.kt
@@ -1,0 +1,13 @@
+package korlibs.bignumber
+
+// Big Integer
+val Long.bi get() = BigInt(this)
+val Int.bi get() = BigInt(this)
+val String.bi get() = BigInt(this)
+fun String.bi(radix: Int) = BigInt(this, radix)
+
+// Big Number
+val Double.bn get() = BigNum("$this")
+val Long.bn get() = BigNum(this.bi, 0)
+val Int.bn get() = BigNum(this.bi, 0)
+val String.bn get() = BigNum(this)
