+++ kbignum/src/commonMain/kotlin/korlibs/bignumber/internal/IntExt.kt
@@ -1,0 +1,13 @@
+package korlibs.bignumber.internal
+
+import kotlin.math.*
+
+// https://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetParallel
+@OptIn(ExperimentalStdlibApi::class)
+internal fun Int.bitCount(): Int = countOneBits()
+
+@OptIn(ExperimentalStdlibApi::class)
+internal fun Int.trailingZeros(): Int = countTrailingZeroBits()
+
+@OptIn(ExperimentalStdlibApi::class)
+internal fun Int.leadingZeros(): Int = countLeadingZeroBits()
