--- kbignum/src/commonTest/kotlin/com/soywiz/kbignum/internal/ProgressionUtilTest.kt
@@ -1,17 +1,0 @@
-package com.soywiz.kbignum.internal
-
-import com.soywiz.kbignum.*
-import kotlin.test.*
-
-class ProgressionUtilTest {
-    @Test
-    fun test() {
-        assertEquals(4.bi, getProgressionLastElement(0.bi, 4.bi, (+1).bi))
-        assertEquals(0.bi, getProgressionLastElement((4).bi, 0.bi, (-1).bi))
-
-        assertEquals((-1).bi, getProgressionLastElement((-4).bi, 0.bi, (+3).bi))
-        assertEquals(3.bi, getProgressionLastElement(0.bi, 4.bi, (+3).bi))
-        assertEquals((1).bi, getProgressionLastElement((4).bi, 0.bi, (-3).bi))
-        assertFailsWith<IllegalArgumentException> { getProgressionLastElement(0.bi, 4.bi, 0.bi) }
-    }
-}
