--- kbignum/src/nativeMain/kotlin/com/soywiz/kbignum/BigIntNative.kt
@@ -1,3 +1,0 @@
-package com.soywiz.kbignum
-
-actual val BigIntNativeFactory: BigIntConstructor = CommonBigInt
