+++ kbignum/src/nativeMain/kotlin/korlibs/bignumber/BigIntNative.kt
@@ -1,0 +1,3 @@
+package korlibs.bignumber
+
+actual val BigIntNativeFactory: BigIntConstructor = CommonBigInt
